requirements.txt
cuda-python
cuda-core